solid effort refactoring gameinfo page
