--- src/components/CombinedScorecard/components/PlayerColumns.jsx
@@ -1,12 +1,0 @@
-// Legacy code migrated to ScorecardRow, deletion pending
-import React from 'react'
-
-export default function PlayerColumns({player}) {
-  return (
-    <tr>
-      <th>{player.jerseyNumber}</th>
-      <th>{player.lastName}</th>
-      <th>{player.defensivePosition.positionAbbr}</th>
-    </tr>
-  )
-}
