+++ src/components/CombinedScorecard/components/ScorecardRow.jsx
@@ -1,7 +1,7 @@
 // This file represents a complete row in the scorecard, with player info on the left, baseball info on the right
 import React from "react";
 import ScoringCell from "./ScoringCell";
-import EmptyCell from "./EmptyCell";
+import EmptyCell from "../../../pages/GameInfo/components/scorecards/EmptyCell";
 import { defensivePositionMapper } from "../../../util/constants";
 
 export default function ScorecardRow({ player, pbp, teamPbp }) {
