+++ src/pages/GameInfo/components/recap/Recap.jsx
