+++ src/pages/GameInfo/components/scorecards/EmptyCell.jsx
