+++ src/pages/GameInfo/components/scorecards/PitchingRecords.jsx
