+++ src/pages/GameInfo/components/scorecards/PitchingRecordsRow.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { defensivePositionMapper } from "../../../util/constants";
+import { defensivePositionMapper } from "../../../../util/constants";
 
 export default function PitcherRow({ pitcher }) {
