+++ src/pages/GameInfo/components/scorecards/Scorecards.jsx
