+++ src/pages/GameInfo/components/scorecards/ScorecardsTeam.jsx
