+++ src/pages/GameInfo/components/scorecards/ScoringRow5.jsx
@@ -1,8 +1,8 @@
 // Legacy code migrated to ScorecardRow, deletion pending
 
 import React from "react";
-import ScoringCell5 from "../../../components/CombinedScorecard/components/ScoringCell5";
-import EmptyCell from "../../../components/CombinedScorecard/components/EmptyCell";
+import ScoringCell5 from "./ScoringCell5";
+import EmptyCell from "./EmptyCell";
 // import {getPlayerIdsPerInning} from "../util/scoringIdentity"
 
 export default function ScoringRow5({ order }) {
